% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fss_utility_functions.R
\name{aggregate_per_location}
\alias{aggregate_per_location}
\title{Aggregate (sum) values over all data streams, and cumulatively over time.}
\usage{
aggregate_per_location(A)
}
\arguments{
\item{A}{An array with three dimensions. Dimensions are: 
\describe{
  \item{Dimension 1}{Time, ordered from most recent to most distant.}
  \item{Dimension 2}{Location, enumerated from 1 and up.}
  \item{Dimension 3}{Data stream, enumerated from 1 and up.}
}}
}
\value{
A matrix with \code{dim(A)[1]} rows and \code{dim(A)[2]} columns.
}
\description{
Aggregate (sum) values over all data streams, and cumulatively over time.
}
\keyword{internal}
