% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_zipstat_over_duration}
\alias{calc_zipstat_over_duration}
\title{Calculate the ZIP window statistic over all durations, for a given zone.}
\usage{
calc_zipstat_over_duration(duration, p, mu, y, maxdur, tol = 0.01)
}
\arguments{
\item{duration}{An integer vector.}

\item{p}{A numeric vector of the given/estimated excess zero probabilities 
corresponding to each count.}

\item{mu}{A numeric vector of the given/estimated Poisson expected value
parameters corresponding to each count. Of same length as \code{p}.}

\item{y}{An integer vector of the observed counts, of same length as 
\code{p}.}

\item{tol}{A scalar between 0 and 1. It is the absolute tolerance criterion
for the estimate of the excess zero indicator; convergence is reached when
two successive elements in the sequence of estimates have an absolute 
difference less than \code{tol}.}
}
\value{
A list with two elements:
\describe{
  \item{duration}{Vector of integers from 1 to \code{maxdur}.}
  \item{statistic}{Numeric vector containing the ZIP statistics corresponding
  to each duration, for the given spatial zone.}
}
}
\description{
This function calculates the zero-inflated Poisson statistic for a given 
spatial zone, for all durations considered.
}
\keyword{internal}

