% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\name{closest_subsets}
\alias{closest_subsets}
\title{Set of increasing sets from left to right of input vector.}
\usage{
closest_subsets(v)
}
\arguments{
\item{v}{A vector. Meant to represent the \eqn{k} nearest neighbors
of a location, the first element being the integer identifier of the 
location itself.}
}
\description{
Returns a set of the increasing sets of the input vector \code{v},
in the sense that the first set contains the first element of \code{v},
the second set the first and second elements of \code{v}, and so on.
}
\keyword{internal}

