% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_utility_functions.R
\name{column_to_int}
\alias{column_to_int}
\title{Replaces a column with integers, its unique and sorted values enumerated.}
\usage{
column_to_int(table, colname)
}
\arguments{
\item{table}{A \code{data.table}.}

\item{colname}{The name of a column in the table. This column should be of
class character or factor.}
}
\value{
The input table \strong{modified}, with the given column replaced by 
   integer values as described.
}
\description{
Takes a \code{data.table} and replaces a single column (preferably) 
containing character or factor values with the enumeration of these values.
The enumeration is done according to the unique and sorted values of the 
original column elements. This function \strong{modifies} the input table.
}
\keyword{internal}

