% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\name{connected_neighbors}
\alias{connected_neighbors}
\title{Find the connected sets for a location and its \eqn{k} nearest neighbors.}
\usage{
connected_neighbors(neighbors, adjacency_matrix)
}
\arguments{
\item{neighbors}{A vector of neighbors to a location, the first element
of the vector being the specific location, and the other elements its 
other nearest neighbors. Locations should be encoded as integers.}

\item{adjacency_matrix}{A boolean matrix, with element \eqn{(i,j)} set 
to TRUE if location \eqn{j} is adjacent to location \eqn{i}.}
}
\value{
Returns a \code{set} of \code{set}s, each set of the latter type 
   containing the location itself and zero or more of its neighbors, if they 
   are connected.
}
\description{
Returns a \code{set} of \code{set}s, each set of the latter type containing
the location itself and zero or more of its neighbors, if they are connected.
}
\keyword{internal}

