% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\name{connected_to}
\alias{connected_to}
\title{Return those elements in the second set which are connected to those in the
first.}
\usage{
connected_to(Z_0, Z_1, adjacency_matrix)
}
\arguments{
\item{Z_0}{A set of locations, given as integers.}

\item{Z_1}{A set of locations, given as integers.}

\item{adjacency_matrix}{A boolean matrix, with element \eqn{(i,j)} set 
to TRUE if location \eqn{j} is adjacent to location \eqn{i}.}
}
\value{
A set, possibly empty, containing those locations in \eqn{Z_1}
        that are connected to any of the locations in \eqn{Z_0}.
}
\description{
Return those elements in the second set \eqn{Z_1} which are connected to 
those in the first set \eqn{Z_0}, according to the adjacency matrix.
}
\keyword{internal}
