% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_negbin.R
\name{convolute_denominator}
\alias{convolute_denominator}
\title{Computes the sum in the increasing score denominator.}
\usage{
convolute_denominator(x, d)
}
\arguments{
\item{x}{A vector of normalized counts summed over a single zone.}

\item{d}{A vector of outbreak durations considered.}
}
\value{
A vector of length \code{length(d)}.
}
\description{
Computes the sum in the increasing score denominator.
}
\keyword{internal}

