% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_utility_functions.R
\name{create_streamset_table}
\alias{create_streamset_table}
\title{Converts a list of stream sets to a \code{data.table} of stream sets and 
streams.}
\usage{
create_streamset_table(streamsets, keys = NULL, offset = 0L)
}
\arguments{
\item{streamsets}{A list of stream sets, elements being vectors of integers
representing different data streams.}

\item{keys}{Character vector of one or more column names which is passed 
to \code{\link[data.table]{setkey}}.}

\item{offset}{An integer to offset the stream set numbering by, in case they
are not named, and you want the numbering to start at \code{offset} + 1.}
}
\description{
Supply a list of stream sets, with each element of the list being a vector
of streams. If the list is named, the output \code{data.table}
will have these names for the stream sets. Else, they will be labeled by 
integers from 1 to the length of the stream set list.
}
\examples{
\dontrun{
create_streamset_table(list(1L, 2L, 1:2))
create_streamset_table(sets::set(sets::set(1L), 
                     sets::set(2L), sets::as.set(1:2)))
create_streamset_table(list(1L, 2L, 1:2), keys = "stream")
create_streamset_table(list(1L, 2L, 1:2), keys = "streamset")
create_streamset_table(list(a = "x", b = "y", c = c("x", "y")))
}
}
\keyword{internal}

