% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_functions.R
\name{cumsum_duration}
\alias{cumsum_duration}
\title{Calculates the cumulative sum of columns over duration, by other columns.}
\usage{
cumsum_duration(table, sumcols, bycols)
}
\arguments{
\item{table}{A \code{data.table} with column \code{duration} and other 
columns matching the other arguments to the function.}

\item{sumcols}{A character vector of the columns to calculate the cumulative
sum for.}

\item{bycols}{A character vector of the columns to calculate the cumulative
sum by (i.e. as the argument \code{by} to a \code{data.table}).}
}
\value{
A \code{data.table} with the cumulative sums over duration.
}
\description{
This function calculates the cumulative sum over the (column) duration for
one or more columns, by some other columns.
}
\keyword{internal}

