% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{df_to_matrix}
\alias{df_to_matrix}
\title{Convert a long data frame to a wide matrix.}
\usage{
df_to_matrix(df, time_col = 1, location_col = 2, value_col = 3)
}
\arguments{
\item{df}{A data frame with at least 3 columns.}

\item{time_col}{Integer or string that specifies the time column.}

\item{location_col}{Integer or string that specifies the location column.}

\item{value_col}{Integer or string that specifies the value column.}
}
\value{
A matrix with time on rows and locations on columns.
}
\description{
Convert a long data frame to a wide matrix, with time along the row dimension
and locations along the column dimension. Values in the matrix could be e.g.
the observed counts or the population.
}
