% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_utility_functions.R
\name{enumerate_character}
\alias{enumerate_character}
\title{Enumerate the unique, sorted values of a column, returned as a list.}
\usage{
enumerate_character(table, colname)
}
\arguments{
\item{table}{A \code{data.table}.}

\item{colname}{The name of a column in the table. This column should be of
class character or factor.}
}
\value{
A list with the sorted and unique values of the column; the original
   values of the column are the list names, and the value for each name is 
   the number for that name (numbering starts at 1).
}
\description{
Enumerate the unique and sorted values of a \code{data.table} column, and 
return the result as a list.
}
\keyword{internal}

