% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{est_eb_zip_zeroindic}
\alias{est_eb_zip_zeroindic}
\title{Calculate the conditional expectation of the structural zero indicator.}
\usage{
est_eb_zip_zeroindic(mu, p, q)
}
\arguments{
\item{mu}{The expected values of the count (which is zero).}

\item{p}{The structural zero probability.}

\item{q}{A scalar greater than or equal to 1; the relative risk.}
}
\value{
A scalar between 0 and 1.
}
\description{
Calculate the conditional expectation of the structural zero indicator for 
the zero-inflated Poisson distribution in the expectation-based ZIP scan
statistic model.
}
\keyword{internal}
