% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{est_zip_relrisk}
\alias{est_zip_relrisk}
\title{Estimate the relative risk for the ZIP distribution.}
\usage{
est_zip_relrisk(y_sum, mu, p, d_hat)
}
\arguments{
\item{y_sum}{A non-negative integer; the sum of the observed counts.}

\item{mu}{A vector of positive scalars; the expected values of the counts.}

\item{p}{A vector of scalars between 0 and 1; the structural zero
probabilities.}

\item{d_hat}{A vector of estimates of the structural zero indicators.}
}
\value{
A scalar; the relative risk.
}
\description{
Estimate the relative risk for the ZIP distribution.
}
\keyword{internal}

