% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_zip.R
\name{estimate_d}
\alias{estimate_d}
\title{Estimate the indicators of excess zeros for a ZIP distribution.}
\usage{
estimate_d(p, mu, y)
}
\arguments{
\item{p}{A numeric vector, each element being the given/estimated 
probability of an excess zero for the corresponding count in \code{y}.}

\item{mu}{A numeric vector, each element being the given/estimated Poisson 
expected value parameter for the corresponding count in \code{y}. Of same 
length as \code{p}.}

\item{y}{An integer vector containing the observed counts. Of same length as 
\code{p}.}
}
\value{
A numeric vector, of same length as the input vector \code{p}.
}
\description{
Given counts and (estimated) Poisson expected value parameters and excess 
zero probabilities, this function estimates the indicator variable for an 
excess zero, for each count.
}
\keyword{internal}

