% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_zip.R
\name{estimate_d_dagger}
\alias{estimate_d_dagger}
\title{Estimate the ZIP excess zero indicators under null hypothesis of no outbreak.}
\usage{
estimate_d_dagger(table)
}
\arguments{
\item{table}{A \code{data.table} with columns \code{p, mu, count}. \code{p}
is the given/estimated probability of an excess zero, and \code{mu} is
the estimated Poisson expected value parameter. \code{count} is the 
observed count.}
}
\value{
The same table, \strong{modified} with a new column \code{ddagger}.
}
\description{
Estimate the excess zero indicator for a zero-inflated Poisson distribution,
under the null hypothesis of no outbreak. The estimate is added as a new 
column to the input table.
}
\keyword{internal}

