% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_struc_zero}
\alias{estimate_struc_zero}
\title{Calculate the conditional expectation of the structural zero indicator.}
\usage{
estimate_struc_zero(mu, p, q)
}
\arguments{
\item{mu}{The expected values of the count (which is zero).}

\item{p}{The structural zero probability.}

\item{q}{A scalar greater than or equal to 1; the relative risk.}
}
\value{
A scalar between 0 and 1.
}
\description{
Calculate the conditional expectation of the structural zero indicator.
}
\keyword{internal}

