% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_estimation.R
\name{estimate_zip_params}
\alias{estimate_zip_params}
\title{Estimate the parameters of a ZIP distribution.}
\usage{
estimate_zip_params(counts, population = NULL)
}
\arguments{
\item{counts}{A matrix of observed counts. Rows indicate time (ordered from 
most recent) and columns indicate locations.}

\item{population}{A matrix or vector of populations for each location 
(optional). If a matrix, should be of the same dimensions as 
\code{counts}. If a vector, should be of the same length as the number of
columns in \code{counts}.}

\item{constant_dim}{An integer. If equal to 1, probabilities are assumed to 
be constant over time but different between locations. If equal to 2, 
probabilities are assumed to vary over time but at each time point be 
equal for all locations.}
}
\value{
A list with two elements:
   \describe{
     \item{baselines}{A matrix of the same dimensions as \code{counts}.}
     \item{probs}{A matrix of the same dimensions as \code{counts}.}
   }
}
\description{
Estimate the ZIP distribution Poisson mean parameters and the structural zero 
probabilities for each location and time point.
}
\keyword{internal}

