% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_zip.R
\name{estimate_zip_relrisk}
\alias{estimate_zip_relrisk}
\title{Estimate the relative risk for an outbreak using a ZIP distribution.}
\usage{
estimate_zip_relrisk(d, mu, y)
}
\arguments{
\item{d}{A vector of indicator variables for whether the corresponding count
in the argument \code{y} is an excess zero or not. Can also be estimates 
of these indicators.}

\item{mu}{A vector of given/estimated Poisson expected value parameters, of 
same length as \code{d}.}

\item{y}{An integer vector of the observed counts, of same length as 
\code{d}.}
}
\value{
A scalar, the estimated relative risk.
}
\description{
Scalar estimate of the relative risk \eqn{q} for zero-inflated Poisson data.
}
\keyword{internal}

