% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fss_exponential_family.R
\name{exponential_lambda}
\alias{exponential_lambda}
\title{Compute the score for an individual exponential observation.}
\usage{
exponential_lambda(c, b)
}
\arguments{
\item{c}{A scalar; the observed value.}

\item{b}{A scalar; the expected value.}
}
\value{
A scalar; the score.
}
\description{
Compute the score for an individual exponential observation.
}
\keyword{internal}
