% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fss_exponential_family.R
\name{exponential_qmax}
\alias{exponential_qmax}
\title{Calculates the largest value for which the exponential score function is
zero.}
\usage{
exponential_qmax(c, b)
}
\arguments{
\item{c}{A scalar; the observed value.}

\item{b}{A scalar; the expected value.}
}
\value{
A scalar greater than or equal to 1.
}
\description{
Calculates the largest value for which the exponential score function is
zero.
}
\keyword{internal}
