% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanstatistics_general.R
\name{extract_mlc}
\alias{extract_mlc}
\title{Extract the most likely cluster (MLC) and the value of the scans statistic.}
\usage{
extract_mlc(table)
}
\arguments{
\item{table}{A \code{data.table} with a column \code{statistic}, which should
correspond to the statistic calculated for each spatial or space-time 
window (given as other columns).}
}
\value{
The row of the input table with the highest value of the column 
   \code{statistic}.
}
\description{
This function extracts the most likely cluster, which is the spatial or 
spatiotemporal window that corresponds to the scan statistic. It also returns
the value of the scan statistc.
}
\keyword{internal}

