% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanstatistics_general.R
\name{extract_scanstatistic}
\alias{extract_scanstatistic}
\title{Extract value of scan statistic from per-window statistics.}
\usage{
extract_scanstatistic(table)
}
\arguments{
\item{table}{A \code{data.table} with a column \code{statistic}, which should
correspond to the statistic calculated for each spatial or space-time 
window (given as other columns).}
}
\value{
The maximum value of the column \code{statistic}.
}
\description{
This function extracts the value of the scan statistic, which is the maximum
of the statistics calculated for each spatial or space-time window.
}
\keyword{internal}

