% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_utility_functions.R
\name{first_keys_match}
\alias{first_keys_match}
\title{Do the first few keys of the supplied table match the supplied keys?}
\usage{
first_keys_match(data_table, keys)
}
\arguments{
\item{data_table}{A \code{data.table}.}

\item{keys}{A character vector.}
}
\value{
\code{TRUE} if keys match, \code{FALSE} otherwise.
}
\description{
Checks if the first few key columns of the supplied \code{data.table} matches
those found in the supplied character vector, in the given order. Boolean 
output.
}
\keyword{internal}

