% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{flipud}
\alias{flipud}
\title{Flip a matrix upside down}
\usage{
flipud(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
A matrix, \code{x} with rows reversed.
}
\description{
Flip a matrix upside down
}
\keyword{internal}
