% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fss_exponential_family.R
\name{gaussian_lambda}
\alias{gaussian_lambda}
\title{Compute the score for an individual Gaussian observation.}
\usage{
gaussian_lambda(c, b, s2)
}
\arguments{
\item{c}{A scalar; the observed value.}

\item{b}{A scalar; the expected value.}

\item{s2}{A scalar; the variance.}
}
\value{
A scalar; the score.
}
\description{
Compute the score for an individual Gaussian observation.
}
\keyword{internal}
