% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_negbin.R
\name{gen_negbin_counts}
\alias{gen_negbin_counts}
\title{Randomly generate and add negative binomial counts to a table.}
\usage{
gen_negbin_counts(table)
}
\arguments{
\item{table}{A \code{data.table} with at least the columns \code{mu} and
\code{theta}. The parameter \eqn{\theta} (theta) is the same as 
\code{size} in \code{\link[stats]{rnbinom}}.}
}
\value{
The same table, with a new column \code{count}.
}
\description{
This function randomly generates counts from a negative binomial distribution
according to the parameters on each row of the input \code{data.table}, and 
adds the counts to a new column \code{count}.
}
\keyword{internal}

