% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_poisson.R
\name{generate_poisson_counts}
\alias{generate_poisson_counts}
\title{Randomly generate and add Poisson counts to a table.}
\usage{
generate_poisson_counts(table)
}
\arguments{
\item{table}{A \code{data.table} with at least the column \code{mu}, 
corresponding to the parameter \code{lambda} in 
\code{\link[stats]{rpois}}.}
}
\description{
This function randomly generates counts from a Poisson distribution according 
to the parameters on each row of the input \code{data.table}, and adds the 
counts to a new column \code{count}.
}
\keyword{internal}

