% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_all_streams}
\alias{get_all_streams}
\title{Extract the sorted names of all data streams from a \code{data.table}.}
\usage{
get_all_streams(d)
}
\arguments{
\item{d}{A \code{data.table} containing a column \code{stream}.}
}
\description{
From a \code{data.table} with a column \code{stream}, extract the unique and
sorted values from this column as a vector.
}
\keyword{internal}

