% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_utility_functions.R
\name{get_column_values}
\alias{get_column_values}
\title{Extract the values of a \code{data.table} column by the column name.}
\usage{
get_column_values(table, colname)
}
\arguments{
\item{table}{A \code{data.table}.}

\item{colname}{The name of a column in the table.}
}
\value{
The values of the column as a vector.
}
\description{
This function extracts the values of a \code{data.table} column as a vector;
the column name is supplied as a character.
}
\keyword{internal}

