% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_utility_functions.R
\name{get_enumerated_character}
\alias{get_enumerated_character}
\title{Get the name (as character) corresponding to the given number, from a list of
enumerated names.}
\usage{
get_enumerated_character(enum_list, number)
}
\arguments{
\item{enum_list}{A list as returned from \code{\link{enumerate_character}}.}

\item{number}{The number of the name you wish to extract.}
}
\description{
Get the name (as character) corresponding to the given number, from a list of
enumerated names.
}
\keyword{internal}

