% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_set}
\alias{get_set}
\title{Get the set with the given index from an implicitly ordered set.}
\usage{
get_set(set_of_sets, index)
}
\arguments{
\item{set_of_sets}{A \code{set} of \code{set}s. The elements of the outer set 
should be ordered, e.g. as it is when the elemets of the sets within it 
are integers.}

\item{index}{The index of the set you wish to be returned.}
}
\description{
Get the set with the given index from an implicitly ordered set.
}
\keyword{internal}

