% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fss_utility_functions.R
\name{get_subset}
\alias{get_subset}
\title{Get either a location or stream subset.}
\usage{
get_subset(lst, subset, d = 3)
}
\arguments{
\item{lst}{A list.}

\item{d}{Integer; either 2 (locations) or 3 (streams).}
}
\value{
A list.
}
\description{
Get either a location or stream subset.
}
\keyword{internal}
