% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_zero_indices}
\alias{get_zero_indices}
\title{Get indices of zero elements in a vector.}
\usage{
get_zero_indices(v)
}
\arguments{
\item{v}{An integer vector.}
}
\value{
A vector with the indices of elements equal to zero in \code{v}.
   Indices start at zero.
}
\description{
Get indices of zero elements in a vector.
}
\keyword{internal}
