% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_utility_functions.R
\name{get_zone}
\alias{get_zone}
\title{Extract a zone from the set of all zones.}
\usage{
get_zone(n, zones)
}
\arguments{
\item{n}{An integer; the number of the zone you wish to retrieve.}

\item{zones}{A list of integer vectors, representing the set of all zones.}
}
\value{
An integer vector.
}
\description{
Extract zone number \eqn{n} from the set of all zones.
}
\examples{
zones <- list(1L, 2L, 3L, 1:2, c(1L, 3L), c(2L, 3L))
get_zone(4, zones)
}

