% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_utility_functions.R
\name{getkeys}
\alias{getkeys}
\title{Get the keys from a data.table.}
\usage{
getkeys(data_table)
}
\arguments{
\item{data_table}{A \code{data.table}.}
}
\value{
NULL if the supplied \code{data.table} has no keys,
   else a character vector containing the keys.
}
\description{
Get the keys from a data.table.
}
\keyword{internal}

