% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_utility_functions.R
\name{haskeys}
\alias{haskeys}
\title{Determine if table has all the given keys.}
\usage{
haskeys(data_table, keys)
}
\arguments{
\item{data_table}{A \code{data.table}.}

\item{keys}{A character vector.}
}
\value{
\code{TRUE} if the table has all the given keys, \code{FALSE} 
   otherwise.
}
\description{
Determine if table has all the given keys.
}
\keyword{internal}

