% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{incomplete_loglihood_term}
\alias{incomplete_loglihood_term}
\title{Calculate a term of the incomplete information loglihood.}
\usage{
incomplete_loglihood_term(y, mu, p, q)
}
\arguments{
\item{y}{A non-negative integer; the observed count.}

\item{mu}{A positive scalar; the expected value of the count.}

\item{p}{A scalar between 0 and 1; the structural zero probability.}

\item{q}{A scalar greater than or equal to 1; the relative risk.}
}
\value{
A non-positive scalar; the loglihood contribution of the 
   observation.
}
\description{
Calculate a term of the incomplete information loglihood for the 
zero-inflated Poisson distribution.
}
\keyword{internal}

