% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{incomplete_zip_loglihood}
\alias{incomplete_zip_loglihood}
\title{Calculate the incomplete information loglihood.}
\usage{
incomplete_zip_loglihood(y, mu, p, q)
}
\arguments{
\item{y}{A non-negative integer vector; the observed counts.}

\item{mu}{A vector of positive scalars; the expected values of the counts.}

\item{p}{A vector of scalars between 0 and 1; the structural zero
probabilities.}

\item{q}{A scalar greater than or equal to 1; the relative risk.}
}
\value{
A non-positive scalar; the incomplete information ZIP loglihood.
}
\description{
Calculate the incomplete information loglihood for the zero-inflated Poisson 
distribution.
}
\keyword{internal}

