% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\name{is_connected}
\alias{is_connected}
\title{Returns TRUE if the neighboring locations are connected to the given 
location, FALSE if not.}
\usage{
is_connected(neighbor_locations, location, adjacency_matrix)
}
\arguments{
\item{neighbor_locations}{A \code{set} of neighboring locations to the given
location; these neighbors do not include the given location itself.}

\item{location}{A location, preferably given as an integer.}

\item{adjacency_matrix}{A boolean matrix, with element \eqn{(i,j)} set 
to TRUE if location \eqn{j} is adjacent to location \eqn{i}.}
}
\value{
Boolean: is the neighbors connected to the given location?
}
\description{
Returns TRUE if the neighboring locations are connected to the given 
location, FALSE if not.
}
\keyword{internal}
