% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{matrix_to_df}
\alias{matrix_to_df}
\title{Convert a matrix to a data frame.}
\usage{
matrix_to_df(mat, name, locations = NULL, times = NULL)
}
\arguments{
\item{mat}{A matrix.}

\item{name}{The name of the third column in the output matrix.}

\item{locations}{If not \code{NULL}, a vector with the names of the 
locations.}

\item{times}{If not \code{NULL}, a vector with the time points. If 
\code{NULL}, the matrix is assumed to be ordered with time point 1 in the 
first row.}
}
\value{
A matrix with columns \code{time, location, name}, where \code{name}
   is specified in the input.
}
\description{
Convert a matrix to a data frame with columns time, location, and one more
containing the elements of the input matrix.
}
\keyword{internal}
