% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_negbin.R
\name{negbin_increasing_score}
\alias{negbin_increasing_score}
\title{Calculate the increasing score for each space-time window.}
\usage{
negbin_increasing_score(table)
}
\arguments{
\item{table}{A \code{data.table} with columns \code{zone, duration, num, 
denom}.}
}
\value{
A \code{data.table} with columns \code{zone, duration, statistic}.
}
\description{
Computes the score statistic for each space-time window, assuming the 
increasing type of outbreak/event/anomaly and either a Poisson or a negative 
binomial distribution for the counts.
}
\keyword{internal}

