% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_negbin.R
\name{negbin_score_terms}
\alias{negbin_score_terms}
\title{Computes the numerator and denominator terms for the NegBin score.}
\usage{
negbin_score_terms(table)
}
\arguments{
\item{table}{A \code{data.table} with columns \code{location, duration, mu,
overdispersion, count}.}
}
\value{
A \code{data.table} with columns \code{location, duration, num, 
   denom}.
}
\description{
This function calculates the terms found in the numerator and denominator 
sums for the ordinary version of the negative binomial scan statistic.
}
\keyword{internal}

