% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NM_geo}
\alias{NM_geo}
\title{Longitude and latitude of New Mexico county seats.}
\format{A data table with 32 rows and 4 variables:
\describe{
  \item{county}{The name of the county. Of class character.}
  \item{seat}{The name of the county seat, i.e. the administrative center or
              seat of government. Of class factor with 32 levels.}
  \item{long}{The longitude of the county seat. Of class numeric.}
  \item{lat}{The latitude of the county seat. Of class numeric.}
}}
\source{
\url{https://en.wikipedia.org/wiki/List_of_counties_in_New_Mexico}
}
\usage{
NM_geo
}
\description{
A dataset containing the longitude and latitude of the county seats of New
Mexico, except for Cibola county.
}
\keyword{datasets}

