% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NM_map}
\alias{NM_map}
\title{Data to plot the counties of New Mexico.}
\format{A data frame with 867 rows and 7 variables:
\describe{
  \item{long}{Numeric; longitude of county polygon corner.}
  \item{lat}{Numeric; latitude of county polygon corner.}
  \item{group}{Numeric; grouping by county.}
  \item{order}{Numeric; order of the polygon corners.}
  \item{region}{Character; region is "new mexico" for all rows.}
  \item{subregion}{Character; the county name (with spaces).}
  \item{county}{Factor; the county name (no spaces).}
}}
\usage{
NM_map
}
\description{
Map data for New Mexico. Was created using \code{ggplot2::map_data}.
}
\keyword{datasets}
