% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NM_popcas}
\alias{NM_popcas}
\title{Population and brain cancer cases in New Mexico counties during 1973--1991.}
\format{A data table with 608 rows and 4 variables:
\describe{
  \item{year}{The year the cases were recorded. Of class integer.}
  \item{county}{The name of the county. Of class factor with 32 levels.}
  \item{population}{The population in that county and year. Of class 
                    numeric.}
  \item{count}{The number of brain cancer cases in that county and year. Of
               class integer.}
}}
\usage{
NM_popcas
}
\description{
A dataset containing the population count and number of brain cancer cases in 
the counties of New Mexico during the years 1973--1991. The population 
numbers are interpolations from the censuses conducted in 1973, 1982, and 
1991. Interpolations were done using a quadratic function of time. Thus the
year-to-year changes are overly smooth but match the census numbers in the 
three years mentioned.
}
\keyword{datasets}

