% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permute_matrix}
\alias{permute_matrix}
\title{Permute the entries of the matrix, preserving row and column marginals.}
\usage{
permute_matrix(A)
}
\arguments{
\item{A}{An integer matrix.}
}
\value{
An integer matrix.
}
\description{
Permute the entries of the matrix, preserving row and column marginals.
}
\keyword{internal}
