% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_poisson.R
\name{poisson_calculations}
\alias{poisson_calculations}
\title{Calculate the expectation-based Poisson statistic for each space-time window.}
\usage{
poisson_calculations(table, zones)
}
\arguments{
\item{table}{A \code{data.table} with columns \code{location, duration, 
count, mu}.}

\item{zones}{A \code{list} or \code{set} of zones, each zone itself a 
set containing one or more locations of those found in \code{table}.}
}
\value{
A \code{data.table} with columns \code{zone, duration, statistic}.
   The column \code{statistic} contains the logarithm of the statistic for 
   each zone-duration combination.
}
\description{
Calculate the expectation-based Poisson statistic for each space-time window,
given the initial data of counts and mus.
}
\keyword{internal}

