% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fss_exponential_family.R
\name{poisson_lambda}
\alias{poisson_lambda}
\title{Compute the score for an individual Poisson observation.}
\usage{
poisson_lambda(c, b)
}
\arguments{
\item{c}{A non-negative integer; the observed count.}

\item{b}{A positive scalar; the expected count.}
}
\value{
A scalar; the score.
}
\description{
Compute the score for an individual Poisson observation.
}
\keyword{internal}
