% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{poisson_loglihood}
\alias{poisson_loglihood}
\title{Calculate the Poisson log-likelihood.}
\usage{
poisson_loglihood(y, mu, q)
}
\arguments{
\item{y}{A non-negative integer vector; the observed counts.}

\item{mu}{A vector of positive scalars; the expected values of the counts.}

\item{q}{A scalar greater than or equal to 1; the relative risk.}
}
\value{
A non-positive scalar; the log-likelihood.
}
\description{
Calculate the log-likelihood for the Poisson distribution.
}
\keyword{internal}
