% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{poisson_lpmf}
\alias{poisson_lpmf}
\title{The log probability mass function of the Poisson distribution.}
\usage{
poisson_lpmf(x, lambda)
}
\arguments{
\item{x}{An integer.}

\item{lambda}{A positive scalar.}
}
\value{
A log-probability.
}
\description{
The log probability mass function of the Poisson distribution.
}
\keyword{internal}

