% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fss_exponential_family.R
\name{poisson_qmax}
\alias{poisson_qmax}
\title{Calculates the largest value for which the Poisson score function is zero.}
\usage{
poisson_qmax(c, b)
}
\arguments{
\item{c}{A non-negative integer; the observed count.}

\item{b}{A positive scalar; the expected count.}
}
\value{
A scalar greater than or equal to 1.
}
\description{
Calculates the largest value for which the Poisson score function is zero.
}
\keyword{internal}
