% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_poisson.R
\name{poisson_relrisk}
\alias{poisson_relrisk}
\title{Add column for relative risk MLE to table with aggregate counts and means.}
\usage{
poisson_relrisk(table)
}
\arguments{
\item{table}{A \code{data.table} with columns \code{zone, duration, count,
mu}. The latter two are the sums of observed counts and estimated expected
value parameters for the locations comprising the zone and up to the given 
duration (i.e. cumulative sum for the duration).}
}
\value{
The same table, with an extra column \code{relrisk}.
}
\description{
This function adds a column for the maximum likelihood estimate for the
relative risk (assumed to be 1 or greater). The table should contain the
aggregates (sums) of the counts and expected value parameters for each zone 
and duration.
}
\keyword{internal}

