% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fss_exponential_family.R
\name{poisson_score}
\alias{poisson_score}
\title{Compute the Poisson score for each priority subset, for each duration.}
\usage{
poisson_score(counts, baselines, priority_indices)
}
\arguments{
\item{counts}{A matrix of observed counts. Rows represent timepoints, ordered
from most recent to most distant. Columns represent e.g. locations or
data streams.}

\item{baselines}{A matrix of expected counts with the same dimensions as
\code{counts}.}

\item{priority_indices}{A matrix of the same size as the input. On each row
(duration), column numbers are given in order of priority.}
}
\value{
A matrix of the same dimension as the input matrices.
}
\description{
Compute the Poisson score for each priority subset, for each duration.
}
\keyword{internal}
