% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_negbin.R
\name{poisson_score_terms}
\alias{poisson_score_terms}
\title{Computes the numerator and denominator terms for the Poisson score.}
\usage{
poisson_score_terms(table)
}
\arguments{
\item{table}{A \code{data.table} with columns \code{location, duration, mu,
count}.}
}
\value{
A \code{data.table} with columns \code{location, duration, num, 
   denom}.
}
\description{
This function calculates the terms found in the numerator and denominator 
sums for the ordinary version of the Poisson scan statistic (Tango et al. 
version).
}
\keyword{internal}

