% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_utility_functions.R
\name{powerset_zones}
\alias{powerset_zones}
\title{Creates a set of all non-empty subsets of the integers from 1 to \eqn{n}.}
\usage{
powerset_zones(n)
}
\arguments{
\item{n}{An integer larger than 0.}
}
\value{
A list of integer vectors.
}
\description{
Creates a list of all \eqn{2^(n-1)} non-empty subsets of the integers from 1 
to \eqn{n}.
}
\keyword{internal}
