% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanstatistics_general.R
\name{print.scanstatistic}
\alias{print.scanstatistic}
\title{Print a scanstatistic object.}
\usage{
\method{print}{scanstatistic}(x, ...)
}
\arguments{
\item{x}{A an object of class \code{scanstatistic}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Prints a scanstatistic object and returns it invisibly.
}
\keyword{internal}

