% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fss_utility_functions.R
\name{prioritize_and_execute}
\alias{prioritize_and_execute}
\title{Order locations accorder to priority, then apply function.}
\usage{
prioritize_and_execute(.f, A, prioritized_locations, ...)
}
\arguments{
\item{.f}{A function taking a vector as first argument.}

\item{A}{A matrix, e.g. containing counts or baselines. Rows represent time 
(ordered from most recent to most distant), columns represent locations 
(numbered from 1 and up).}

\item{...}{Further arguments passed to \code{.f}.}
}
\value{
A matrix of the same dimension as \code{A}.
}
\description{
Order locations accorder to priority, then apply function.
}
\keyword{internal}
