% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{remove_intlist_duplicates}
\alias{remove_intlist_duplicates}
\title{Remove duplicates from a list of integer vectors.}
\usage{
remove_intlist_duplicates(x)
}
\arguments{
\item{x}{A list of integer vectors.}
}
\value{
A list of integer vectors; the same as input \code{x} but with 
   duplicates removed. The returned list is not named.
}
\description{
Remove duplicates from a list of integer vectors.
}
\keyword{internal}

