% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fss_utility_functions.R
\name{reorder_rows}
\alias{reorder_rows}
\title{Reorder rows by priority.}
\usage{
reorder_rows(A, priority_indices)
}
\arguments{
\item{A}{A matrix, containing e.g. counts or baselines. Rows represent time 
(ordered from most recent to most distant), columns represent e.g. 
locations or data streams (numbered from 1 and up).}

\item{priority_indices}{An integer matrix as output by 
\code{\link{prioritize_cols}}.}
}
\value{
An integer matrix of the same dimension as \code{A}.
}
\description{
Reorder each row in the input matrix \code{A} by the column indices found
in the corresponding row of the matrix \code{priority_indices}.
}
\keyword{internal}
