% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scan_bayes_negbin_cpp}
\alias{scan_bayes_negbin_cpp}
\title{Calculate the "Bayesian Spatial Scan Statistic" by Neill et al. (2006).}
\usage{
scan_bayes_negbin_cpp(counts, baselines, zones, zone_lengths, outbreak_prob,
  alpha_null, beta_null, alpha_alt, beta_alt, m_values, m_probs)
}
\arguments{
\item{counts}{An integer matrix (most recent timepoint in first row).}

\item{baselines}{A matrix with positive entries (most recent timepoint in 
first row).}

\item{zones}{An integer vector (all zones concatenated; locations indexed 
from 0 and up).}

\item{zone_lengths}{An integer vector.}

\item{outbreak_prob}{A scalar; the probability of an outbreak (at any time,
any place).}

\item{alpha_null}{A scalar; the shape parameter for the gamma distribution
under the null hypothesis of no anomaly.}

\item{beta_null}{A scalar; the scale parameter for the gamma distribution
under the null hypothesis of no anomaly.}

\item{alpha_alt}{A scalar; the shape parameter for the gamma distribution
under the alternative hypothesis of an anomaly.}

\item{beta_alt}{A scalar; the scale parameter for the gamma distribution
under the alternative hypothesis of an anomaly.}

\item{m_values}{A vector of possible values for the increase in the mean
(and variance) of an anomalous count.}

\item{m_probs}{A vector of the prior probabilities of each value in 
\code{m_values}.}
}
\value{
A list with elements \code{observed} and \code{simulated}, each 
   being a data frame with columns:
   \describe{
     \item{zone}{The top-scoring zone (spatial component of MLC).}
     \item{duration}{The corresponding duration (time-length of MLC).}
     \item{score}{The value of the loglihood ratio statistic (the scan
                  statistic).}
   }
}
\description{
Calculate the "Bayesian Spatial Scan Statistic" by Neill et al. (2006),
adapted to a spatio-temporal setting. The scan statistic assumes that,
given the relative risk, the data follows a Poisson distribution. The 
relative risk is in turn assigned a Gamma distribution prior, yielding a 
negative binomial marginal distribution for the counts.
}
\keyword{internal}
