% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_eb_poisson.R
\name{scan_eb_poisson}
\alias{scan_eb_poisson}
\title{Calculate the expectation-based Poisson scan statistic.}
\usage{
scan_eb_poisson(counts, zones, baselines = NULL, population = NULL,
  n_mcsim = 0, gumbel = TRUE, max_only = FALSE, rel_tol = 0.001)
}
\arguments{
\item{counts}{A matrix of observed counts. Rows indicate time and are ordered
from least recent (row 1) to most recent (row \code{nrow(counts)}). 
Columns indicate locations, numbered from 1 and up.}

\item{zones}{A list of integer vectors. Each vector corresponds to a single
zone; its elements are the numbers of the locations in that zone.}

\item{baselines}{A matrix of the same dimensions as \code{counts}. Holds the
Poisson mean parameter of the ZIP distribution for each observed count.}

\item{population}{A matrix or vector of populations for each location. Only
needed if \code{baselines} and \code{probs} are to be estimated and you 
want to account for the different populations in each location (and time).
If a matrix, should be of the same dimensions as \code{counts}. If a 
vector, should be of the same length as the number of columns in 
\code{counts}.}

\item{n_mcsim}{A non-negative integer; the number of replicate scan 
statistics to generate in order to calculate a P-value.}

\item{gumbel}{Boolean. If \code{TRUE}, and \code{n_mcsim > 0}, then a Gumbel
distribution is fit to the replicate scan statistics and a \eqn{p}-value
for the observed scan statistic is calculated using the fitted 
distribution. If \code{FALSE} and \code{n_mcsim > 0}, the empirical
\eqn{p}-value is calculated instead, defined as 1 + the number of 
replicate scan statistics that were larger than the observed scan 
statistic, divided by \code{1 + n_mcsim}.}

\item{max_only}{Boolean. If \code{FALSE} (default) the log-likelihood ratio
statistic for each zone and duration is returned. If \code{TRUE}, only the
largest such statistic (i.e. the scan statistic) is returned, along with
the corresponding zone and duration.}

\item{rel_tol}{A positive scalar. If the relative change in the incomplete
information likelihood is less than this value, then the EM algorithm is
deemed to have converged.}
}
\description{
Calculate the expectation-based Poisson scan statistic.
}
\examples{
\dontrun{
set.seed(1)
# Create location coordinates, calculate nearest neighbors, and create zones
geo <- matrix(rnorm(100), 50, 2)
knn_mat <- coords_to_knn(geo, 15)
zones <- knn_zones(knn_mat)

# Simulate data
baselines <- matrix(rexp(50, 1/5), 5, 50)
counts <- rpois(1, baselines, probs)

# Inject outbreak/event/anomaly
ob_dur <- 1:3
ob_zone <- zones[[10]]
counts[ob_dur, ob_zone] <- rpois(
  1, 2 * baselines[ob_dur, ob_zone])
res <- scan_eb_poisson(counts = counts, 
                       zones = zones,
                       baselines = baselines, 
                       n_mcsim = 100,
                       gumbel = TRUE,
                       max_only = FALSE,
                       rel_tol = 1e-3)
}
}
\keyword{internal}

