% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scan_eb_poisson_cpp_max}
\alias{scan_eb_poisson_cpp_max}
\title{Calculate the highest-value EB Poisson loglihood ratio statistic.}
\usage{
scan_eb_poisson_cpp_max(counts, agg_baselines, zones, zone_lengths)
}
\arguments{
\item{counts}{A matrix of non-negative integers; the observed counts. Rows
indicate time, ordered from most recent (row 1) to least recent. Columns
indicate locations; the locations are numbered from 1 and up.}

\item{agg_baselines}{A matrix of positive scalars; the expected values of 
the counts, cumulatively summed over locations (columns). Of the same 
dimensions as \code{counts}.}

\item{zones}{An integer vector containing the zones, stored one after
another. Each zone is found using the elements of the parameter
\code{zone_lengths}. For example, if the first element of
\code{zone_lengths} is 5, then the first 5 elements of \code{zones}
make up the first zone. If the second element of \code{zone_lengths} is
2, then elements 6 and 7 of \code{zones} make up the second zone, and so
on. Note that the zones are numbered from 0 and up in the input, but
from 1 and up in the output.}

\item{zone_lengths}{An integer vector holding the number of locations in
each zone.}
}
\value{
A data frame with five columns:
   \describe{
     \item{zone}{The top-scoring zone (spatial component of MLC).}
     \item{duration}{The corresponding duration (time-length of MLC).}
     \item{score}{The value of the loglihood ratio statistic (the scan
                  statistic).}
     \item{relrisk}{The estimated relative risk.}
   }
}
\description{
Calculate the expectation-based Poisson loglihood ratio statistic for each 
zone and duration, but only keep the zone and duration with the highest 
value (the MLC). The estimate of the relative risks inside and outside the 
MLC is also calculated and returned.
}
\keyword{internal}
