% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scan_pb_zip_cpp}
\alias{scan_pb_zip_cpp}
\title{Calculate the highest-value PB ZIP loglihood ratio statistic.}
\usage{
scan_pb_zip_cpp(counts, pop, zones, zone_lengths, num_locs, num_zones, max_dur,
  rel_tol, store_everything)
}
\arguments{
\item{counts}{matrix}

\item{pop}{matrix}

\item{zones}{list of integer vectors}

\item{zone_lengths}{vector}

\item{num_locs}{int}

\item{num_zones}{int}

\item{max_dur}{int}

\item{rel_tol}{double}

\item{store_everything}{boolean}
}
\value{
A data frame with five columns:
   \describe{
     \item{zone}{The top-scoring zone (spatial component of MLC).}
     \item{duration}{The corresponding duration (time-length of MLC).}
     \item{score}{The value of the loglihood ratio statistic (the scan
                  statistic).}
     \item{relrisk}{The estimated relative risk.}
     \item{n_iter}{The number of iterations performed by the EM algorithm.}
   }
}
\description{
Calculate the population-based ZIP loglihood ratio statistic for each zone
and duration, but only keep the zone and duration with the highest value
(the MLC). The estimate of the relative risk is also calculated, along with
the number of iterations the EM algorithm performed.
}
\keyword{internal}
