% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanstatistics_general.R
\name{scanstatistic_object}
\alias{scanstatistic_object}
\title{Creates an S3 object of class scanstatistic.}
\usage{
scanstatistic_object(observed, simulated, details)
}
\arguments{
\item{observed}{A data table containing columns location, duration, 
statistic, and possibly others.}

\item{simulated}{A numeric vector of replicated scan statistics.}

\item{details}{A list containing details about the data and scan statistic.}
}
\description{
Creates an S3 object of class scanstatistic.
}
\keyword{internal}

