% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanstatistics.R
\docType{package}
\name{scanstatistics-package}
\alias{scanstatistics}
\alias{scanstatistics-package}
\title{scanstatistics: Space-time anomaly detection using scan statistics.}
\description{
The scanstatistics package provides two categories of important functions:
data preparation functions, and the scan statistics themselves.
}
\section{Data preparation functions}{

These functions prepare your data for use. In particular, it helps you 
define the \emph{zones} which will be considered by the scan statistics.
}

\section{Scan statistics}{

These are the functions used for space-time anomaly detection. Scan statistic
functions for univariate space-time data have a name that begins with 
\code{scan_} and functions for multivariate space-time data have a name that
begins with \code{mscan_}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/promerpr/scanstatistics}
  \item Report bugs at \url{https://github.com/promerpr/scanstatistics/issues}
}

}
\author{
\strong{Maintainer}: Paul Romer Present \email{paul.romerpresent@fastmail.fm} [contributor]

Authors:
\itemize{
  \item Benjamin Allévius \email{benjak@math.su.se}
}

}
