% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{score_zip}
\alias{score_zip}
\title{Calculate the ZIP loglihood ratio statistic and the relative risk.}
\usage{
score_zip(y, mu, p, rel_tol = 0.01)
}
\arguments{
\item{y}{A non-negative integer vector; the observed counts.}

\item{mu}{A vector of positive scalars; the expected values of the counts.}

\item{p}{A vector of scalars between 0 and 1; the structural zero
probabilities.}

\item{rel_tol}{The absolute convergence criterion.}

\item{d_hat}{A vector of estimates of the structural zero indicators.}
}
\value{
A list with three elements:
   \enumerate{
     \item The loglihood ratio statistic.
     \item The estimated relative risk.
     \item The number of iterations of the EM algorithm performed.
   }
}
\description{
Calculate the ZIP loglihood ratio statistic and the relative risk.
}
\keyword{internal}

