% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_negbin.R
\name{score_zone_sums}
\alias{score_zone_sums}
\title{Sums the numerator and denominator terms over all locations in each zone.}
\usage{
score_zone_sums(table, zones)
}
\arguments{
\item{table}{A \code{data.table} with columns \code{location, duration, num,
denom}; the output from \code{\link{negbin_score_terms}}.}

\item{zones}{A \code{set} of zones, each zone itself being a \code{set}
containing locations. Locations should be encoded as integers.}
}
\value{
A \code{data.table} with columns \code{zone, duration, num, denom}.
}
\description{
Computes the sum of the numerator and denominator terms over all locations in
each zone, as part of the score calculation.
}
\keyword{internal}

