% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fss_main_functions.R
\name{subset_aggregation_FF}
\alias{subset_aggregation_FF}
\title{Fast Subset Aggregation over both locations and data streams.}
\usage{
subset_aggregation_FF(args, score_fun = poisson_score,
  priority_fun = poisson_priority, R = 50, rel_tol = 0.01)
}
\arguments{
\item{args}{A list of matrices:
   \describe{
     \item{counts}{Required. A matrix of counts. Rows indicate time, ordered 
                   from most recent to most distant. Columns indicate e.g. 
                   locations or data streams, enumerated from 1 and up.}
     \item{baselines}{Required. A matrix of expected counts. Dimensions are 
                      as for \code{counts}.}
     \item{penalties}{Optional. A matrix of penalty terms. Dimensions are as
                      for \code{counts}.}
     \item{...}{Optional. Futher matrices with parameters}
}}

\item{score_fun}{A function taking matrix arguments, all of the
same dimension, and returning a matrix or vector of that dimension.}

\item{priority_fun}{A function taking matrix arguments, all of the
same dimension, and returning a matrix or vector of that dimension.}

\item{R}{The number of random restarts.}

\item{rel_tol}{The relative tolerance criterion. If the current score divided
by the previous score, minus one, is less than this number then the 
algorithm is deemed to have converged.}
}
\value{
A list containing the most likely cluster (MLC), having the following 
   elements:
   \describe{
     \item{score}{The score of the MLC.}
     \item{duration}{The duration of the MLC, i.e. how many time periods from
                     the present into the past the MLC stretches.}
     \item{locations}{The locations contained in the MLC.}
     \item{streams}{The data streams contained in the MLC.}
     \item{random_restarts}{The number of random restarts performed.}
     \item{avg_iter_to_conv}{The average number of iterations it took to
                             reach convergence.}
   }
}
\description{
Compute the most likely cluster (MLC) with the Subset Aggregation method by
Neill et al. (2013) through fast optimization over subsets of locations and 
subsets of streams.
}
\details{
Note: algorithm not quite as in Neill et al. (2013) since the 
   randomly chosen subset of streams is the same for all time windows.
}
\references{
Neill, Daniel B., Edward McFowland, and Huanian Zheng (2013). \emph{Fast 
   subset scan for multivariate event detection}. Statistics in Medicine 
   32 (13), pp. 2185-2208.
}
\keyword{internal}

