% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fss_main_functions.R
\name{subset_aggregation_FN_NF}
\alias{subset_aggregation_FN_NF}
\title{Compute the most likely cluster using the FN/NF Subset Aggregation algorithm.}
\usage{
subset_aggregation_FN_NF(args, score_fun = poisson_score,
  priority_fun = poisson_priority, algorithm = "FN")
}
\arguments{
\item{args}{A list of matrices:
   \describe{
     \item{counts}{Required. A matrix of counts. Rows indicate time, ordered 
                   from most recent to most distant. Columns indicate e.g. 
                   locations or data streams, enumerated from 1 and up.}
     \item{baselines}{Required. A matrix of expected counts. Dimensions are 
                      as for \code{counts}.}
     \item{penalties}{Optional. A matrix of penalty terms. Dimensions are as
                      for \code{counts}.}
     \item{...}{Optional. Futher matrices with parameters}
}}

\item{score_fun}{A function taking matrix arguments, all of the
same dimension, and returning a matrix or vector of that dimension.}

\item{priority_fun}{A function taking matrix arguments, all of the
same dimension, and returning a matrix or vector of that dimension.}

\item{d}{An integer: \code{d=2} means sums are taken over locations (the 
second dimension) and \code{d=3} means sums are taken over data streams.
Other values of \code{d} should not be used.}
}
\description{
Compute the most likely cluster (MLC) with the Subset Aggregation method by
Neill et al. (2013), either through fast optimization over subsets of 
locations and naive optimization over subsets of streams (FN), or through
naive optimization over subsets of locations and fast optimization over 
subsets of streams (NF).
}
\references{
Neill, Daniel B., Edward McFowland, and Huanian Zheng (2013). \emph{Fast 
   subset scan for multivariate event detection}. Statistics in Medicine 
   32 (13), pp. 2185-2208.
}
\keyword{internal}

