% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fss_utility_functions.R
\name{sum_over_subset}
\alias{sum_over_subset}
\title{For each array in a list, sum over dimension \code{d}.}
\usage{
sum_over_subset(lst, subset, d = 3)
}
\arguments{
\item{lst}{List containing 3-dimensional arrays. The second dimension should
represent locations, and the third dimension should represent data 
streams.}

\item{subset}{An integer vector holding a subset of indices.}

\item{d}{An integer; the array dimension to sum over.}
}
\value{
A list containing matrices. This list has the same names as the input
   list.
}
\description{
For each array in a list, sum over dimension \code{d}.
}
\keyword{internal}
