% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fss_utility_functions.R
\name{sum_reorder_sum}
\alias{sum_reorder_sum}
\title{Cumulatively sum, order by priority, then sum again.}
\usage{
sum_reorder_sum(A, priority_indices)
}
\arguments{
\item{A}{A numeric or integer matrix.}

\item{priority_indices}{An integer matrix as output by 
\code{\link{prioritize_cols}}.}
}
\description{
For a given matrix, form a new matrix by cumulatively summing over its 
columns. Reorder the row contents in this matrix by the corresponding indices 
in the given priority matrix. Then cumulatively sum over each row, yielding
a new matrix of the same size as the input matrix.
}
\keyword{internal}
