% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{times_and_durations}
\alias{times_and_durations}
\title{Creates a \code{data.table} with columns \code{time} and \code{duration}.}
\usage{
times_and_durations(d)
}
\arguments{
\item{d}{A \code{data.table} containing at least a column \code{time}, which 
is sortable. For example, could be POSIXct dates.}
}
\value{
A new \code{data.table}, containing columns \code{time} (key column)
   and \code{duration}.
}
\description{
Creates a \code{data.table} with columns \code{time} and \code{duration},
in which the column \code{time} corresponds to the (unique) times in the 
input table. In the output, the most recent time given a duration of 1, the 
second most recent time given a duration of 2, and so on.
}
\keyword{internal}

