% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanstatistics_general.R
\name{validate_colnames}
\alias{validate_colnames}
\title{Check that the input table has the right columns. Raises error if not.}
\usage{
validate_colnames(table, col_names)
}
\arguments{
\item{table}{A data.table passed to a scan statistic function.}

\item{col_names}{A character vector of column names required.}
}
\description{
Check that the input table has the right columns. Raises error if not.
}
\keyword{internal}

