% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanstatistics_general.R
\name{validate_scan}
\alias{validate_scan}
\title{Check that input to scanstatistic function is valid.}
\usage{
validate_scan(table, zones, col_names)
}
\arguments{
\item{table}{A data.table passed to a scan statistic function.}

\item{zones}{Should be a list of integer or factor vectors.}

\item{col_names}{A character vector of column names; these columns in the
table must not have any missing values.}
}
\description{
Check that input to scanstatistic function is valid.
}
\keyword{internal}

