% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanstatistics_general.R
\name{validate_zones}
\alias{validate_zones}
\title{Check that the zones argument is a list of integer or factor vectors.}
\usage{
validate_zones(zones)
}
\arguments{
\item{zones}{Should be a list of integer or factor vectors.}
}
\description{
Check that the zones argument is a list of integer or factor vectors.
}
\keyword{internal}

