% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{zip_loglihood}
\alias{zip_loglihood}
\title{Calculate the ZIP log-likelihood.}
\usage{
zip_loglihood(y, mu, p, q)
}
\arguments{
\item{y}{A vector of non-negative integers; the observed counts.}

\item{mu}{A vector of positive scalars; the expected values of the counts.}

\item{p}{A vector of scalars between 0 and 1; the structural zero
probabilities.}

\item{q}{A scalar greater than or equal to 1; the relative risk.}
}
\value{
A non-positive scalar; the incomplete information ZIP loglihood.
}
\description{
Calculate the (incomplete information) log-likelihood for the zero-inflated 
Poisson distribution.
}
\keyword{internal}

