% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{zip_lpmf}
\alias{zip_lpmf}
\title{The log probability mass function of the ZIP distribution.}
\usage{
zip_lpmf(y, mu, p)
}
\arguments{
\item{y}{A non-negative integer; the observed count.}

\item{mu}{A positive scalar; the expected value of the count.}

\item{p}{A scalar between 0 and 1; the structural zero probability.}
}
\value{
A non-positive scalar; the loglihood contribution of the 
   observation.
}
\description{
The log probability mass function of the Poisson distribution, evaluated at
\code{x}.
}
\keyword{internal}

