% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_zip.R
\name{zip_statistic}
\alias{zip_statistic}
\title{Calculates the ZIP statistic for each space-time window.}
\usage{
zip_statistic(table, ...)
}
\arguments{
\item{table}{A \code{data.table} with columns \code{zone, location, duration, 
p, mu, count}.}

\item{...}{Any of the following named parameters:
\describe{
  \item{maxdur}{As in \code{\link{calc_zipstat_over_duration}}.}
  \item{d_init}{As in \code{link{window_zip_statistic}}.}
  \item{tol}{As in \code{link{window_zip_statistic}}.}
}}
}
\value{
A \code{data.table} with columns \code{zone, duration, statistic}.
}
\description{
Calculates the zero-inflated Poisson statistic for each space-time window,
using the EM algorithm.
}
\keyword{internal}

