% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_functions.R
\name{zone_sum}
\alias{zone_sum}
\title{Sum columns over all location in each zone, for each duration.}
\usage{
zone_sum(table, sumcols)
}
\arguments{
\item{table}{A \code{data.table} with columns \code{zone, location, duration} 
and those given in the argument \code{sumcols}.}

\item{sumcols}{Character vector of column names, the columns to be summed 
over each zone and duration.}
}
\description{
Sum columns over all location in each zone, for each duration.
}
\keyword{internal}

