% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NM_geo}
\alias{NM_geo}
\title{Longitude and latitude of New Mexico county seats.}
\format{A data frame with 33 rows and 7 variables:
\describe{
  \item{county}{Factor; the counties of New Mexico (no spaces).}
  \item{seat}{Character; the name of the county seat, i.e. the administrative 
              center or seat of government.}
  \item{area(km2)}{Numeric; the area in square kilometers of each county.}
  \item{seat_long}{Numeric; the longitude of the county seat.}
  \item{seat_lat}{Numeric; the latitude of the county seat.}
  \item{center_long}{Numeric; the longitude of the geographical center of
                     the county.}
  \item{center_lat}{Numeric; the latitude of the geographical center of the 
                    county.}
}}
\source{
\url{https://en.wikipedia.org/wiki/List_of_counties_in_New_Mexico}
}
\usage{
NM_geo
}
\description{
A dataset containing the longitude and latitude of the county seats of New
Mexico, except for Cibola county.
}
\keyword{datasets}
