% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NM_map}
\alias{NM_map}
\title{Data to plot the counties of New Mexico.}
\format{A data table with 867 rows and 7 variables:
\describe{
  \item{long}{Longitude of county polygon corner. Of class numeric.}
  \item{lat}{Latitude of county polygon corner. Of class numeric.}
  \item{group}{Grouping by county. Of class numeric.}
  \item{order}{Order of the polygon corners. Of class integer.}
  \item{region}{Region is "new mexico" for all rows. Of class character.}
  \item{subregion}{The county name (with spaces). Of class character.}
  \item{county}{The county name (no spaces). Of class factor with 33 levels,
                the county Cibola places last.}
}}
\usage{
NM_map
}
\description{
Map data for New Mexico. Was created using \code{ggplot2::map_data}.
}
\keyword{datasets}

