% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{poisson_lpmf}
\alias{poisson_lpmf}
\title{The log probability mass function of the Poisson distribution.}
\usage{
poisson_lpmf(y, mu)
}
\arguments{
\item{y}{An integer.}

\item{mu}{A positive scalar.}
}
\value{
A log-probability.
}
\description{
The log probability mass function of the Poisson distribution, evaluated at
\code{y}.
}
\keyword{internal}

