% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_pb_poisson.R
\name{scan_pb_poisson}
\alias{scan_pb_poisson}
\title{Calculate the population-based Poisson scan statistic.}
\usage{
scan_pb_poisson(counts, zones, baselines = NULL, population = NULL,
  n_mcsim = 0, gumbel = TRUE, max_only = FALSE)
}
\arguments{
\item{counts}{A matrix of observed counts. Rows indicate time and are ordered
from least recent (row 1) to most recent (row \code{nrow(counts)}).
Columns indicate locations, numbered from 1 and up.}

\item{zones}{A list of integer vectors. Each vector corresponds to a single
zone; its elements are the numbers of the locations in that zone.}

\item{baselines}{A matrix of the same dimensions as \code{counts}. Holds the
Poisson mean parameter of the ZIP distribution for each observed count.
Note that the sum of all elements in \code{baselines} must equal the sum
of all elements in \code{counts}. If baselines is not supplied, it will be
estimated.}

\item{population}{A matrix or vector of populations for each location. Only
needed if \code{baselines} are to be estimated and you want to account for 
the different populations in each location (and time).
If a matrix, should be of the same dimensions as \code{counts}. If a
vector, should be of the same length as the number of columns in
\code{counts}.}

\item{n_mcsim}{A non-negative integer; the number of replicate scan
statistics to generate in order to calculate a P-value.}

\item{gumbel}{Boolean. If \code{TRUE} and \code{n_mcsim > 0}, then a Gumbel
distribution is fit to the replicate scan statistics and a \eqn{p}-value
for the observed scan statistic is calculated using the fitted
distribution. If \code{FALSE} and \code{n_mcsim > 0}, the empirical
\eqn{p}-value is calculated instead, defined as 1 + the number of
replicate scan statistics that were larger than the observed scan
statistic, divided by \code{1 + n_mcsim}.}

\item{max_only}{Boolean. If \code{FALSE} (default) the log-likelihood ratio
statistic for each zone and duration is returned. If \code{TRUE}, only the
largest such statistic (i.e. the scan statistic) is returned, along with
the corresponding zone and duration.}
}
\description{
Calculate the population-based Poisson scan statistic.
}
\examples{
\dontrun{
set.seed(1)
# Create location coordinates, calculate nearest neighbors, and create zones
geo <- matrix(rnorm(100), 50, 2)
knn_mat <- coords_to_knn(geo, 15)
zones <- knn_zones(knn_mat)

# Simulate data
population <- 100 + matrix(2 * rt(50 * 5, df = 2), 5, 50)
baselines <- matrix(rexp(50 * 5, 1/5), 5, 50)
counts <- matrix(rpois(prod(dim(baselines)), as.vector(baselines)),
                 nrow(baselines), ncol(baselines))

# Inject outbreak/event/anomaly
ob_dur <- 1:3
ob_zone <- zones[[10]]
counts[ob_dur, ob_zone] <- rpois(
  1, 2 * baselines[ob_dur, ob_zone])
res <- scan_pb_poisson(counts = counts,
                       zones = zones,
                       n_mcsim = 100,
                       gumbel = TRUE,
                       max_only = FALSE)
}
}
\keyword{internal}

